+++ shell.nix
@@ -1,8 +1,8 @@
 let
     nixpkgs = builtins.fetchGit {
-        name = "nixos-23.11";
+        name = "nixos-24.05";
         url = "https://github.com/nixos/nixpkgs";
-        ref = "refs/heads/nixos-23.11";
+        ref = "refs/heads/nixos-24.05";
     };
 
     nixpkgs-unstable = builtins.fetchGit {
@@ -23,6 +23,9 @@ in pkgs.mkShell {
         gcc
         cmake
         pkg-config
+
+        xdelta
+        libwebp
     ];
 
     buildInputs = with pkgs; [