+++ src/background.rs
@@ -1,3 +1,5 @@
+use std::process::Command;
+
 use anime_launcher_sdk::anime_game_core::installer::downloader::Downloader;
 use anime_launcher_sdk::anime_game_core::minreq;
 
@@ -10,30 +12,46 @@ pub struct Background {
 }
 
 pub fn get_uri() -> String {
-    let uri = concat!("https://sdk-os-static.", "ho", "yo", "verse", ".com/hk4e_global/mdk/launcher/api/content?filter_adv=true&key=gcStgarh&launcher_id=10&language=");
+    let lang = crate::i18n::get_lang();
 
-    uri.to_owned() + &crate::i18n::format_lang(&crate::i18n::get_lang())
+    if lang.language == unic_langid::langid!("zh-cn").language {
+        concat!("https://hyp-api.", "mi", "ho", "yo", ".com/hyp/hyp-connect/api/getAllGameBasicInfo?launcher_id=jGHBHlcOq1").to_owned()
+    }
+
+    else {
+        let uri = concat!("https://sg-hyp-api.", "ho", "yo", "verse", ".com/hyp/hyp-connect/api/getAllGameBasicInfo?launcher_id=VYTpXlbWo8&language=");
+
+        uri.to_owned() + &crate::i18n::format_lang(&lang)
+    }
 }
 
 #[cached::proc_macro::cached(result)]
 pub fn get_background_info() -> anyhow::Result<Background> {
     let json = serde_json::from_slice::<serde_json::Value>(minreq::get(get_uri()).send()?.as_bytes())?;
 
-    let uri = match json["data"]["adv"]["background"].as_str() {
-        Some(uri) => uri.to_owned(),
-        None => anyhow::bail!("Failed to get background picture uri")
-    };
+    let uri = json["data"]["game_info_list"].as_array()
+        .ok_or_else(|| anyhow::anyhow!("Failed to list games in the backgrounds API"))?
+        .iter()
+        .find(|game| {
+            match game["game"]["biz"].as_str() {
+                Some(biz) => biz.starts_with("hk4e_"),
+                _ => false
+            }
+        })
+        .ok_or_else(|| anyhow::anyhow!("Failed to find the game in the backgrounds API"))?["backgrounds"]
+        .as_array()
+        .and_then(|backgrounds| backgrounds.iter().next())
+        .and_then(|background| background["background"]["url"].as_str())
+        .ok_or_else(|| anyhow::anyhow!("Failed to get background picture url"))?
+        .to_string();
 
-    // This API field contains wrong md5 hash, but file's name
-    // from the uri above actually contains correct one, so
-    // I parse and use it few lines below
-
-    /*let hash = match json["data"]["adv"]["bg_checksum"].as_str() {
-        Some(uri) => uri.to_owned(),
-        None => anyhow::bail!("Failed to get background picture checksum")
-    };*/
-
-    let hash = uri.split('/').last().unwrap_or_default().split('_').next().unwrap_or_default().to_owned();
+    let hash = uri.split('/')
+        .last()
+        .unwrap_or_default()
+        .split('_')
+        .next()
+        .unwrap_or_default()
+        .to_owned();
 
     Ok(Background {
         uri,
@@ -46,22 +64,46 @@ pub fn download_background() -> anyhow::Result<()> {
 
     let info = get_background_info()?;
 
+    let mut download_image = true;
+
     if crate::BACKGROUND_FILE.exists() {
         let hash = Md5::digest(std::fs::read(crate::BACKGROUND_FILE.as_path())?);
 
         if format!("{:x}", hash).to_lowercase() == info.hash {
             tracing::debug!("Background picture is already downloaded. Skipping");
 
-            return Ok(());
+            download_image = false;
+
+            if crate::BACKGROUND_PRIMARY_FILE.exists() {
+                tracing::debug!("Background picture is already patched. Skipping");
+
+                return Ok(());
+            }
         }
     }
 
-    let mut downloader = Downloader::new(info.uri)?;
+    if download_image {
+        let mut downloader = Downloader::new(&info.uri)?;
 
-    downloader.continue_downloading = false;
+        downloader.continue_downloading = false;
 
-    if let Err(err) = downloader.download(crate::BACKGROUND_FILE.as_path(), |_, _| {}) {
-        anyhow::bail!(err);
+        if let Err(err) = downloader.download(crate::BACKGROUND_FILE.as_path(), |_, _| {}) {
+            anyhow::bail!(err);
+        }
+    }
+
+    // Workaround for GTK weakness
+    if info.uri.ends_with(".webp") {
+        Command::new("dwebp")
+            .arg(crate::BACKGROUND_FILE.as_path())
+            .arg("-o")
+            .arg(crate::BACKGROUND_PRIMARY_FILE.as_path())
+            .spawn()?
+            .wait_with_output()?;
+    }
+
+    else {
+        std::fs::copy(crate::BACKGROUND_FILE.as_path(), crate::BACKGROUND_PRIMARY_FILE.as_path())?;
     }
 
     Ok(())